refactor(lotto): 상수 변경
- 상수화 필요한 부분 수정
- 클래스 이름 변경 : Constants.Error 클래스를 Errors로 변경

+++ src/main/java/lotto/controller/GameController.java
@@ -83,7 +83,7 @@ public class GameController {
         return result;
     }
 
-    private void getRate(long purchaseMoney, long reward) {
+    private void getRate(int purchaseMoney, int reward) {
         Rate rate = new Rate(purchaseMoney, reward);
         outputView.println(rate);
     }
